Fix for bonemeal exploit
Placing bone meal on grass grows long grass & plants without taking away the bone meal, as it's placed on the floor instead. This is just a quick fix, if you want a proper fix you need to cancel the event if they're sneaking and then only place the stack while sneaking, or something like that.

(Not tested, did this directly on GitHub. Should work fine though.)